Kane-AI-iOS/requirements.txt
Appium-Python-Client==4.0.1
boto3==1.38.17
botocore==1.38.17
selenium==4.17.2
pydantic==2.8.2
pydantic_core==2.20.1
lxml==5.2.1
Pillow==11.0.0
requests==2.28.2
httpx==0.27.0
python-json-logger==2.0.7
pyotp==2.9.0
